requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-WTF==1.2.1
WTForms==3.1.2
Werkzeug==3.0.4
google-cloud-bigquery==3.25.0
requests==2.32.3
python-dotenv==1.0.1
pytz==2024.1
gunicorn==23.0.0
